# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/investor/thesis": {
      "get": {
        "operationId": "InvestorController_getThesis",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "post": {
        "operationId": "InvestorController_createOrUpdateThesis",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThesisDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "delete": {
        "operationId": "InvestorController_deleteThesis",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/scoring": {
      "get": {
        "operationId": "InvestorController_getScoringWeights",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "put": {
        "operationId": "InvestorController_updateScoringWeights",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScoringWeightsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/matches": {
      "get": {
        "operationId": "InvestorController_getMatches",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "default": 1,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "exclusiveMinimum": true,
              "default": 20,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "integer"
            }
          },
          {
            "name": "isSaved",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/matches/{startupId}": {
      "get": {
        "operationId": "InvestorController_getMatchDetails",
        "parameters": [
          {
            "name": "startupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/matches/{startupId}/save": {
      "patch": {
        "operationId": "InvestorController_toggleSaved",
        "parameters": [
          {
            "name": "startupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/matches/{matchId}/status": {
      "patch": {
        "operationId": "InvestorController_updateMatchStatus",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/notes": {
      "post": {
        "operationId": "InvestorController_createNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "get": {
        "operationId": "InvestorController_getAllNotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/notes/{startupId}": {
      "get": {
        "operationId": "InvestorController_getNotes",
        "parameters": [
          {
            "name": "startupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/notes/{noteId}": {
      "patch": {
        "operationId": "InvestorController_updateNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "delete": {
        "operationId": "InvestorController_deleteNote",
        "parameters": [
          {
            "name": "noteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/portfolio": {
      "post": {
        "operationId": "InvestorController_addToPortfolio",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPortfolioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      },
      "get": {
        "operationId": "InvestorController_getPortfolio",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
...
    "/investor/team/invite": {
      "post": {
        "operationId": "InvestorController_createTeamInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/team/invite/{id}": {
      "delete": {
        "operationId": "InvestorController_cancelInvite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/team/member/{id}": {
      "delete": {
        "operationId": "InvestorController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investor"
        ]
      }
    },
    "/investor/team/join/{inviteCode}": {
      "post": {
        "operationId": "InvestorTeamPublicController_acceptInvite",
        "parameters": [
          {
            "name": "inviteCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "InvestorTeamPublic"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS Boilerplate API",
    "description": "Production-ready NestJS boilerplate with auth, queues, storage, and email",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "CreateThesisDto": {
        "type": "object",
        "properties": {
          "industries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "checkSizeMin": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "checkSizeMax": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "geographicFocus": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mustHaveFeatures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dealBreakers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "maxLength": 5000
          }
        }
      },
      "UpdateScoringWeightsDto": {
        "type": "object",
        "properties": {
          "marketWeight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "teamWeight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "productWeight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "tractionWeight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "financialsWeight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "marketWeight",
          "teamWeight",
          "productWeight",
          "tractionWeight",
          "financialsWeight"
        ]
      },
      "UpdateMatchStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "reviewing",
              "engaged",
              "closed",
              "passed"
            ]
          },
          "passReason": {
            "type": "string",
            "maxLength": 500
          },
          "passNotes": {
            "type": "string",
            "maxLength": 5000
          },
          "investmentAmount": {
            "type": "number",
            "exclusiveMinimum": true,
            "minimum": 0
          },
          "investmentCurrency": {
            "type": "string",
            "maxLength": 10
          },
          "investmentDate": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "investmentNotes": {
            "type": "string",
            "maxLength": 5000
          },
          "meetingRequested": {
            "type": "boolean"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateNoteDto": {
        "type": "object",
        "properties": {
          "startupId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "isPinned": {
            "type": "boolean"
          }
        },
        "required": [
          "startupId",
          "content"
        ]
      },
      "UpdateNoteDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "isPinned": {
            "type": "boolean"
          }
        }
      },
      "AddPortfolioDto": {
        "type": "object",
        "properties": {
          "startupId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "dealSize": {
            "type": "integer",
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0
          },
          "dealStage": {
            "type": "string",
            "maxLength": 100
          },
          "investedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "startupId"
        ]
      },
      "CreateTeamInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "role": {
            "type": "string",
            "enum": [
              "member",
              "admin"
            ]
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
...
    }
  }
}